+++ components/images/StaticImage.tsx
