+++ components/navigation/TopToolbar.tsx
@@ -1,5 +1,5 @@
 import { View, StyleSheet, Text } from 'react-native';
-import { StaticImage } from '@/components/StaticImage';
+import { StaticImage } from '@/components/images/StaticImage';
 import SimpleLineIcons from '@expo/vector-icons/SimpleLineIcons';
 import SettingsMenu from '../menus/SettingsMenu';
 import NotificationsMenu from '../menus/NotificationsMenu';
